--- cortex/.ipynb_checkpoints/requirements-checkpoint.txt
@@ -1,4 +1,0 @@
-gensim==3.8.1
-psycopg2-binary==2.8.4
-
-
