+++ cortex/exported_data/letterboxd/riley/ratings.csv
@@ -1,0 +1,58 @@
+Date,Name,Year,Letterboxd URI,Rating
+2020-02-26,Suspiria,2018,https://letterboxd.com/film/suspiria-2018/,2.5
+2020-02-26,Very Ralph,2019,https://letterboxd.com/film/very-ralph/,5
+2020-02-26,Enter the Void,2009,https://letterboxd.com/film/enter-the-void/,3.5
+2020-02-26,Dogtooth,2009,https://letterboxd.com/film/dogtooth/,3
+2020-02-26,Prisoners,2013,https://letterboxd.com/film/prisoners/,5
+2020-02-26,Under the Skin,2013,https://letterboxd.com/film/under-the-skin-2013/,3
+2020-02-26,Melancholia,2011,https://letterboxd.com/film/melancholia/,3
+2020-02-26,Beyond the Black Rainbow,2010,https://letterboxd.com/film/beyond-the-black-rainbow/,2.5
+2020-02-26,I Saw the Devil,2010,https://letterboxd.com/film/i-saw-the-devil/,3.5
+2020-02-26,Once Upon a Time Veronica,2012,https://letterboxd.com/film/once-upon-a-time-veronica/,3.5
+2020-02-26,Kill List,2011,https://letterboxd.com/film/kill-list/,4.5
+2020-02-26,Night Moves,2013,https://letterboxd.com/film/night-moves-2013/,4
+2020-02-26,Upstream Color,2013,https://letterboxd.com/film/upstream-color/,1
+2020-02-26,The Babadook,2014,https://letterboxd.com/film/the-babadook/,2.5
+2020-02-26,Bone Tomahawk,2015,https://letterboxd.com/film/bone-tomahawk/,4.5
+2020-02-26,Dogville,2003,https://letterboxd.com/film/dogville/,5
+2020-02-26,Annihilation,2018,https://letterboxd.com/film/annihilation/,4.5
+2020-02-26,Coherence,2013,https://letterboxd.com/film/coherence/,3
+2020-02-26,It Follows,2014,https://letterboxd.com/film/it-follows/,4.5
+2020-02-26,The Blackcoat's Daughter,2015,https://letterboxd.com/film/the-blackcoats-daughter/,3
+2020-02-26,Spring,2014,https://letterboxd.com/film/spring-2014/,3
+2020-02-26,Deadwood Park,2007,https://letterboxd.com/film/deadwood-park/,5
+2020-02-26,The House That Jack Built,2018,https://letterboxd.com/film/the-house-that-jack-built-2018/,4
+2020-02-26,Victoria,2015,https://letterboxd.com/film/victoria-2015/,3.5
+2020-02-26,The Witch,2015,https://letterboxd.com/film/the-witch-2015/,5
+2020-02-26,Zodiac,2007,https://letterboxd.com/film/zodiac/,3.5
+2020-02-26,Braid,2018,https://letterboxd.com/film/braid-2018/,1.5
+2020-02-26,Vazante,2017,https://letterboxd.com/film/vazante/,1
+2020-02-26,In the Tall Grass,2019,https://letterboxd.com/film/in-the-tall-grass-2019/,3.5
+2020-02-26,A Dark Song,2016,https://letterboxd.com/film/a-dark-song/,4
+2020-02-26,The Witcher Documentary,2017,https://letterboxd.com/film/the-witcher-documentary/,3.5
+2020-02-26,The Wailing,2016,https://letterboxd.com/film/the-wailing/,3.5
+2020-02-26,Mindhunter - The Making of a Serial Killer,2018,https://letterboxd.com/film/mindhunter-the-making-of-a-serial-killer/,0.5
+2020-02-26,Beast,2017,https://letterboxd.com/film/beast-2017/,4
+2020-02-26,Under the Silver Lake,2018,https://letterboxd.com/film/under-the-silver-lake/,2.5
+2020-02-26,The Killing of a Sacred Deer,2017,https://letterboxd.com/film/the-killing-of-a-sacred-deer/,5
+2020-02-26,Dark,2018,https://letterboxd.com/film/dark-2018/,3.5
+2020-02-26,The Sinner,2017,https://letterboxd.com/film/the-sinner-2017/,4.5
+2020-02-26,The Hole in the Ground,2019,https://letterboxd.com/film/the-hole-in-the-ground-2019/,3.5
+2020-02-26,A Quiet Place,2018,https://letterboxd.com/film/a-quiet-place-2018/,2.5
+2020-02-26,The Rain,2007,https://letterboxd.com/film/the-rain/,3.5
+2020-02-26,Parasite,2019,https://letterboxd.com/film/parasite-2019/,5
+2020-02-26,White Chamber,2018,https://letterboxd.com/film/white-chamber/,0.5
+2020-02-26,Mandy,2018,https://letterboxd.com/film/mandy-2018/,4.5
+2020-02-26,Watchmen,2019,https://letterboxd.com/film/watchmen-2019/,1
+2020-02-26,Joker,2019,https://letterboxd.com/film/joker-2019/,5
+2020-02-26,Nothing Really Happens,2017,https://letterboxd.com/film/nothing-really-happens/,3.5
+2020-02-26,Chernobyl,2019,https://letterboxd.com/film/chernobyl/,5
+2020-02-26,The Perfection,2018,https://letterboxd.com/film/the-perfection/,0.5
+2020-02-26,Hereditary,2018,https://letterboxd.com/film/hereditary/,4
+2020-02-26,The Lighthouse,2019,https://letterboxd.com/film/the-lighthouse-2019/,3.5
+2020-02-26,Earthquake Bird,2019,https://letterboxd.com/film/earthquake-bird/,3.5
+2020-02-26,Climax,2018,https://letterboxd.com/film/climax-2018/,3
+2020-02-26,Antichrist,2009,https://letterboxd.com/film/antichrist/,5
+2020-02-26,Midsommar,2019,https://letterboxd.com/film/midsommar/,2
+2020-02-26,The Society,2015,https://letterboxd.com/film/the-society/,5
+2020-02-26,Back Roads,2018,https://letterboxd.com/film/back-roads-2018/,4
