+++ cortex/requirements.txt
@@ -1,4 +1,4 @@
 gensim==3.8.1
 psycopg2-binary==2.8.4
-
+python-dotenv==0.11.0
